% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_funcs.R
\name{ag_trend}
\alias{ag_trend}
\title{Summarize aggregation results}
\usage{
ag_trend(x, timeframe, ci.prob = 0.95)
}
\arguments{
\item{x}{An aggregated abundance object from \code{ag_abund}}

\item{timeframe}{A 2-vector giving the time frames for the calculated growth trends}

\item{ci.prob}{Confidence (credible) interval value}
}
\description{
Summarize aggregation results
}
