% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ssl_nimble.R
\name{fit_ssl_nimble}
\alias{fit_ssl_nimble}
\title{Fit agTrend SSL model using MCMC with NIMBLE}
\usage{
fit_ssl_nimble(
  x,
  niter = 110000,
  nburnin = 10000,
  thin = 10,
  debug = FALSE,
  ...
)
}
\arguments{
\item{x}{A named list generated from the \code{prep_for_nimble} that prepares the raw data}

\item{niter}{Number of iterations for the MCMC, this includes the burnin and is prior to thinning.}

\item{nburnin}{Number of burnin iterations. These will be discarded from \code{niter} before returning results.}

\item{thin}{Thinning of the MCMC sample for storage. Total number of retuned samples is \code{(niter-nburnin)/thin}.}

\item{debug}{Jump into the fucntion for debugging.}

\item{...}{Additional arguments.}
}
\description{
Fit agTrend SSL model using MCMC with NIMBLE
}
\author{
Devin S. Johnson
}
