% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_nimble_2.R
\name{prep_for_nimble}
\alias{prep_for_nimble}
\title{Create data list for NIMBLE analysis}
\usage{
prep_for_nimble(
  x,
  timeframe,
  model.cuts = c(5, 10),
  max.factor = 3,
  debug = FALSE
)
}
\arguments{
\item{x}{Data containing the Steller sea lion counts and sites}

\item{timeframe}{A 2-vector containing the start and years for the modeling effort}

\item{model.cuts}{A 2-vector specifying the numbers of positive counts necessary for each model type.
E.g., \code{model.cuts=c(5,10)} implies that a constant model (over years)
is used for sites with <=5 positive counts, a linear model is used for the number of positive counts in
(5, 10], and a low-rank Gaussian process is used for the number of positive counts >10.}

\item{max.factor}{Maximum value for imputed counts calculated as \code{max.factor * max(counts)}.
Currently this is not implemented, so, it is ignored}

\item{debug}{Logical. Should debug mode be entered. Probably just for Devin to use.}
}
\description{
Create a named list with data structures necessary to run the MCMC in \code{NIMBLE}
}
\author{
Devin S. Johnson
}
