% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_funcs.R
\name{summary_ag}
\alias{summary_ag}
\title{Summarize aggregation results}
\usage{
summary_ag(x, ci.prob = 0.95, smooth = TRUE)
}
\arguments{
\item{x}{An aggregated abundance object from \code{agg_abund}}

\item{ci.prob}{Confidence (credible) interval value}

\item{smooth}{Logical. Should smoothing of piecewise credible intervals be used. This is primarily used
for plotting. No statistical inference is made from smoothed intervals.}
}
\description{
Summarize aggregation results
}
